requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==26.1.0
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
ConfigArgParse==1.7.5
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
datasets==4.8.3
dictdiffer==0.9.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==1.1.0
dvc==3.67.0
dvc-data==3.18.3
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.135.1
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
funcy==2.0
gevent==25.9.1
geventhttpclient==2.3.9
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
great_expectations==1.15.1
greenlet==3.3.2
gto==1.9.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.7.1
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kombu==5.6.2
locust==2.43.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mistune==3.2.0
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
omegaconf==2.3.0
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
scmrepo==3.6.1
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
simple-websocket==1.1.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
sqltrie==0.11.2
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
vine==5.1.0
voluptuous==0.16.0
waitress==3.0.2
wcwidth==0.6.0
websocket-client==1.9.0
Werkzeug==3.1.6
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zc.lockfile==4.0
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
